Added exceptions

+++ Recruiter.Application/Features/Candidates/Handlers/Commands/DeleteCandidateCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using MediatR;
 using Recruiter.Application.Contracts.Persistence;
+using Recruiter.Application.Exceptions;
 using Recruiter.Application.Features.Candidates.Requests.Commands;
+using Recruiter.Domain;
 
 namespace Recruiter.Application.Features.Candidates.Handlers.Commands;
 
@@ -19,6 +21,8 @@ public class DeleteCandidateCommandHandler : IRequestHandler<DeleteCandidateComm
     public async Task<Unit> Handle(DeleteCandidateCommand request, CancellationToken cancellationToken)
     {
         var candidate = await _candidateRepository.Get(request.Id);
+        if (candidate == null)
+            throw new NotFoundException(nameof(Candidate), request.Id);
         await _candidateRepository.Delete(candidate);
         return Unit.Value;
     }
